Skip the cleanup process if the test agent does not have the permission to delete the test items created during the running of test cases.

+++ TestSuites/FileServer/src/SMB2Model/TestSuite/SMB2ModelTestAssemblyCleanup.cs
@@ -1,11 +1,7 @@
 ﻿using Microsoft.Protocols.TestSuites.FileSharing.SMB2Model.Adapter;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using System;
-using System.Collections.Generic;
 using System.IO;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace Microsoft.Protocols.TestSuites.FileSharing.SMB2Model.TestSuite
 {
@@ -15,11 +11,22 @@ namespace Microsoft.Protocols.TestSuites.FileSharing.SMB2Model.TestSuite
         [AssemblyCleanup]
         public static void AssemblyCleanup()
         {
+            var skipCleanup = false;
+
             foreach (var directory in ModelManagedAdapterBase.AllTestDirectories)
             {
+                if (skipCleanup)
+                {
+                    break;
+                }
+
                 try
                 {
                     Directory.Delete(directory);
+                }
+                catch (UnauthorizedAccessException)
+                {
+                    skipCleanup = true;
                 }
                 catch
                 {
@@ -29,9 +36,18 @@ namespace Microsoft.Protocols.TestSuites.FileSharing.SMB2Model.TestSuite
 
             foreach (var file in ModelManagedAdapterBase.AllTestFiles)
             {
+                if (skipCleanup)
+                {
+                    break;
+                }
+
                 try
                 {
                     File.Delete(file);
+                }
+                catch (UnauthorizedAccessException)
+                {
+                    skipCleanup = true;
                 }
                 catch
                 {
